move metadata tests to failing mocha
